requirements.txt
albumentations==0.4.6
torch==1.12.1
torchvision==0.13.1
Pillow==7.1.2
numpy==1.21.6
